[DL] Optimize address space zero lookup (NFC)
Information for pointer size/alignment/etc is queried a lot, but
currently fairly slow. It uses a binary search, which is
particularly bad the the most common query (address space zero).

Use a simple linear scan instead, which will immediately return
for address space zero, and perform better than the binary search
as long as the number of address spaces is not too large. Even the
AMDGPU backend only registers six address spaces.