[SCEV] Fix and validate ValueExprMap/ExprValueMap consistency
This adds validation for consistency of ValueExprMap and ExprValueMap, and fixes encountered issues. The main one is that addrec construction directly wrote to ValueExprMap without updating ExprValueMap. The adjustment in forgetSymbolicName() is to drop the old value from the map, rather than relying on the fact that it will get overwritten.