Restrict equality propagation for pointers
Reviving https://reviews.llvm.org/D143129

This patch does the following:

Adds the following functions:
- replaceDominatedUsesWithIf() that takes a callback.

- canReplacePointersIfEqual(...) returns true if the underlying object
is the same, and for null and const dereferencable pointer replacements.

- canReplacePointersIfEqualInUse(...) returns true for the above as well
as if the use is in icmp/ptrtoint or phi/selects feeding into them.

Updates GVN using the functions above so that the pointer replacements
are only made using the above API.

Change-Id: I4927ea452734458be028854ef0e5cbcd81955910

Fix typos, remove some uses of undef/poison

Change-Id: I7d0ddd8e6d2bab90de72d2c3f691cfbea9c3e257

Use DataLayout as function argument

Change-Id: Ia70757984f307a0be2d0ef9141325be71653c321