Extern function promotion
Promotion of extern functions is still safe so long as the external
interface remains the same. This commit adds the ability to split an otherwise
argpromotable function into a wrapper function that maintains the
external interface, and a wrappee that gets argpromoted.

In most compilation modes, this doesn't make sense; many extern
functions are called from outside the module, so the split just adds a
pointless layer of indirection. Therefore, we enable this splitting only for
ThinLTO, where the wrapper is actually inlinable.