[InstCombine] Make indexed compare fold GEP source type independent
The indexed compare fold converts comparisons of GEPs with same
(indirect) base into comparisons of offset. Currently, it only
supports GEPs with the same source element type.

This change makes the transform operate on offsets instead, which
removes the type dependence. To keep closer to the scope of the
original implementation, this keeps the limitation that we should
only have at most one variable index per GEP.