[DomTree] Remove unnecessary domtree level check in SemiNCA (NFC)
runSemiNCA() currently checks that the ReverseChildren are below
MinLevel in the DT, which is used when performing incremental
updates.

However, ReverseChildren is populated during runDFS with only
the predecessors that are part of that DFS walk, which will itself
be level limited in the relevant cases. As such, I don't believe
that this should be checked during runSemiNCA().