Add vectorize-s115.ll
Determine Dst and Src overlapping by SCEV of Src and Dist

Consider the following case:

  for (int j = 0; j < 256; j++)    // Loop j
    for (int i = j+1; i < 256; i++)// Loop i
      a[i] -= aa[j][i] * a[j];

Given that SCEV of &a[j] is {@a,+,4}<Loop j>, a[j] will be treated as scalar
when vectorizing Loop i. If the accessing size of a[j] <= Dist(a[j], a[i]),
there is no overlapped and can be vectorized.

In this case, accessing size of a[j] is 4 byte(float) and Dist(a[j], a[i])
is {4,+,4} which bring the minimum distance as 4