Improve applyLoopGuards to suppport Mul
Improve applyLoopGuards to preserve divisibility information
of SCEVMul Expressions.
It does so by preserving the information that a SCEVMul with a SCEVConstant operand divides by this constant.
It also generalized the assumption cache divisibility propagation to non-SCEVUnkinowns.
For example if:

TC = TC1 * TC2;
__builtin_assume(TC % 8 == 0);

We now propagate the information that SCEVMul divides by 8, by rewriting it to ((TC1 * TC2) \ 8) * 8

This fixes #82367.