InstSimplify: increase shufflevector test coverage
Add examples of patterns that can be simplified, but are currently not.
This patch serves as a pre-commit test for #48960.

InstSimplify/test: use poison instead of undef

InstSimplify/shufflevector: add more tests

InstSimplify/shufflevector: remove redundant tests

InstSimplify/shufflevector: add multiuse test

InstSimplify/shufflevector: add negative bitcast test

InstSimplify/shufflevector: add negative cast test

InstSimplify/shufflevector: add identity tests

InstSimplify: lookthru casts, binops in folding shuffles

foldIdentityShuffles() currently suffers from the deficiency that it
cannot look through casts or binary operators, and can only fold
sequences of shuffles without any intervening instructions. Fix this
deficiency by looking through chains of casts that preserve the number
of elements in the vector, and chains of binary operators with identical
operands. The way we do this is by saving the last instruction in the
reduction chain as the RootVec, and rewriting the operands of the first
instruction in the chain.

This patch also has a requirement on increasing the MaxRecursionLimit,
and the AMDGPU and LoopVectorize test changes are due to that.

Future work could simplify the case of binary operands with different
operands, and the multi-use case.

Alive2 proofs follow.

  @fold_lookthrough_cast:
    https://alive2.llvm.org/ce/z/YFKQGX
  @fold_lookthrough_binop_same_operands:
    https://alive2.llvm.org/ce/z/gf3WgB
  @fold_lookthrough_cast_chain:
    https://alive2.llvm.org/ce/z/MT9ApH
  @fold_lookthrough_binop_chain:
    https://alive2.llvm.org/ce/z/WV8HRn
  @fold_lookthrough_cast_binop_chain:
    https://alive2.llvm.org/ce/z/CXj-hh

InstSimplify: address review; minor fixes