Support scaled vscale as step in howFarToZero
VF * vscale is the canonical step for a scalably vectorized loop, and
LFTR canonicalizes to NE loop tests, so having our trip count logic be
unable to compute trip counts for such loops is unfortunate.

The existing code needed minimal generalization to handle non-constant
strides.  The tricky cases to be sure we handle correctly are: zero,
and -1 (due to the special case of abs(-1) being non-positive).

This patch does the full generalization in terms of code structure, but
only lets it kick in when the non-constant is a (C * vscale) node.  I
did some quick investigation, and it seems the context free non-zero,
and sign checks are basically never disproved for arbitrary scales.
I think we have alternate tactics available for these, but I'm going
to return to that in a separate patch.