Use cond from header as upper bound on IV in exit BB.
For loops, we can use the condition in the loop header as upper bound on
the compared induction in the unique exit block, if it exists. This can
be done even if there are multiple in-loop edges to the unique exit
block, as any other exit may only exit earlier.

More generally, we could add the OR of all exit conditions to the exit,
but that's a possible future extension.

Fixes https://github.com/llvm/llvm-project/issues/90417.

!fixup properly handle non-monotonically increasing case.