[EarlyCSE] Insert unreachable for !noundef undef load
This extends #96639 for mem2reg to earlycse. If we try to replace
a load with !noundef metadata by undef, insert a non-terminator
unreachable.

I've added a handleLoadOfUndef() helper for this purpose, which is
shared between mem2reg and earlycse (and in the future also gvn
etc). This also provides a place to experiment with replacing
uninit loads with freeze poison.