Push more users to worklist for simplifyUsers
Instead of considering only users that are inside the loop,
consider all the users with parent dominated by the loop header.

Add simple tests

Default to enable full traversal (it shoud be changed before merge)

typo

Update tests

Update test for LoopDeletion/LoopVectorize

Fix eliminateTrunc

As we can now go outside the loop, the IV user is not necessarily
invariant. Using L->isLoopInvariant to determine which operand is the
IV user is wrong in this context.