[SROA] Avoid expensive isComplete() call (NFC)
https://github.com/llvm/llvm-project/pull/83381 introduced a call
to PHINode::isComplete() in Mem2Reg, which is O(n^2) in the number
of predecessors, resulting in pathological compile-time regressions
for cases with many predecessors.

I believe this was intended as a compile-time optimization in the
first place, so just drop the check, and always try to transfer
the attribute.