Speed up ReachingDefAnalysis (NFC)
This patch speeds up ReachingDefAnalysis without changing the
functionality.

Without this patch, ReachingDefAnalysis maintains a three-dimensional
vector to keep track of reaching definitions (i.e. instructions) for
each basic block and each reg unit.  Quite literally:

  MBBReachingDefs[MBBNumber][Unit]

is a vector of definitions, where the innermost vector is implemented
with TinyPtrVector.  The problem is that for many pairs of MBBNumber
and Unit, we have more than one definition, causing TinyPtrVector to
allocate SmallVector, accounting for 1.91% of heap allocations during
the compilation of X86ISelLowering.cpp.ll, a .ll version of
X86ISelLowering.cpp.

This patch replaces the three-dimensional vector with a custom
multimap implemented on a single instance of
std::vector<std::pair<int, int>>.  The vector contains a collection of
singly-linked lists of reaching definitions for all possible pairs of
MBBNumber and Unit.

With this patch, the instruction count and cycle count go down by
0.24% and 0.82%, respectively, for the compilation of
X86ISelLowering.cpp.ll on a stabilized x86 machine (the frequency
govenor set to performance and hyperthreading disabled).  The number
of heap allocations goes down by 2.20%.