ConstraintElim: teach fact-transfer about samesign
When the samesign flag is present on an icmp, we can transfer all the
facts on the unsigned system to the signed system, and vice-versa: we do
this by specializing transferToOtherSystem when samesign is present.
Supporting samesign in ConstraintElimination has necessitated adding
extra information to ConditionFacts, as the information would otherwise
be lost.