Improve Sema diagnostic performance for __builtin_counted_by_ref
Implement the sema checks with a placeholder. We then check for that
placeholder in all of the places we care to emit a diagnostic.

Don't use a placeholder. Instead, place diagnostics in the situations we want to disallow.

Remove removed placeholder type.

Remove unused code.

Extract check for __builtin_counted_by_ref CallExpr into a helper method.

Reformat.

Turn into a const method.

Use helper function.