Perform loop versioning to use memcpy
LoopIdiomRecognize has given up on processing loops if an alias can
exist for the target load and store. This patch introduces loop
versioning within the LoopIdiomRecognize and improves the memcpy
replacement in such cases. Currently, loop versioning is only performed
for the loop that is dedicated to transferring content from a load to a
store is supported. This limitatin prevents the code size from growing
up.

Related to #50892