Add initial nodes to the Worklist in VectorCombine
This tries to mirror how InstructionWorklist is used in InstCombine, adding the
nodes initially to a list that is added in reverse order to the Worklist.  The
general order should be the same, the main advantage of this is that as node
are initially processed, when altered the New and Old instructions are visited
immediately, helping remove Old instructions as they are replaced, helping
other combines work without hitting OneUse checks.