Port all SME routines to RuntimeLibcalls
This updates everywhere we emit/check an SME routines to use
RuntimeLibcalls to get the function name and calling convention.

Note: RuntimeLibcallEmitter had some issues with emitting non-unique
variable names for sets of libcalls, so tweaked the output to avoid
the need for variables.

Rebase: Use RuntimeLibcalls for EmitEntryPStateSM

Avoid TLI.getSupportedLibcallImpl

Rewrite to check a (much smaller) list of libcalls