Limit recursion in getRangeRef for PHI nodes.
Restrict PHI nodes that getRangeRef is allowed to recursively examine so
we don't need a "visited" set.  And fix createSCEVIter so it creates all
the relevant SCEV nodes before getRangeRef tries to examine them.

The tests that are affected have induction variables that aren't
AddRecs.  (Other cases are theoretically affected, but don't seem to
show up in our tests.)