[InstCombine] Allow freezing multiple operands
InstCombine tries to convert `freeze(inst(op))` to `inst(freeze(op))`.
Currently, this is limited to the case where a single operand needs
to be frozen, and all other operands are guaranteed non-poison.

This patch allows the transform even if multiple operands need to be
frozen. The existing limitation makes sure that we do not increase
the total number of freezes, but it also means that that we may
fail to eliminate freezes (via poison flag dropping) and may prevent
optimizations (as analysis generally can't look past freeze).
Overall, I believe that aggressively pushing freezes upwards is more
beneficial than harmful.

This is the middle-end version of #145939 in DAGCombine (which is
currently reverted for SDAG-specific reasons).

Fix test name

reset variable names

Do single pass recursive propagation