SimplifyCFG: Add tests for switch-replacements
Currently, some switch optimizations are only possible, if lookup table
support is enabled. However, some of these optimizations are no lookup
tables at all, but a series of simple instructions like mul, add
or shift instructions.

These tests will show how a future change will enable these other kinds
of optimization, even if lookup tables are not supported.

SimplifyCFG: Add `isLookupTable` method

Check if the SwitchReplacement is a lookup table or not.

SimplifyCFG: Delay check for switch-to-lookup option

The option prevents any switch replacements from happening earlier in
the compilation pipeline, even though it should only apply to lookup
tables.

Delay the check until we are sure that we would create a lookup table.
If not, continue with the replacement.

SimplifyCFG: Delay check for LUT support in target

Only check the target support for LUTs if we would create one.
Otherwise, proceed with the optimization.

SimplifyCFG: Delay check for "no-jump-tables" attribute

Allow the replacement of switches with anything other than a lookup
table, even if the "no-jump-table" attribute is set.

SimplifyCFG: Delay creation of bitmaps

Early bitmap creation can, similar to lookup tables,
cause missed optimizations because they are more difficult to analyze.

Delay the creation until later in the pipeline, when the option
`switch-to-lookup` is set in in SimplifyCFG.