Improve `foldDeadPhiWeb` compile time
The foldDeadPhiWeb function identifies and removes small dead PHI webs, it bails
out if the web size exceeds threshold (16).

In the current implementation, when there is a phi node has large number of
users that most of them are phi nodes, we still push them on the `Stack` even
if the numbers of phi nodes user already exceed the threshold. It is unnecessary
and may consume too much time.

This patch checks the early stop condition when we push an unvisited phi node
on the `Stack`.
With this change, the wall duration of total instcombine pass decreased from
523,649.276 ms to 30,239.399 ms in an our internal case.

clang format

fix...